+++ http/internal/logic/user/changepasswordlogic.go
@@ -29,7 +29,7 @@ func NewChangePasswordLogic(ctx context.Context, svcCtx *svc.ServiceContext) *Ch
 func (l *ChangePasswordLogic) ChangePassword(req *types.ChangePasswordRequest) (resp *types.ChangePasswordResponse, err error) {
 	uidjson, _ := l.ctx.Value("UID").(json.Number)
 	uid, _ := uidjson.Int64()
-	user, _ := l.svcCtx.UserModel.FindOne(l.ctx, uint64(uid))
+	user, _ := l.svcCtx.UserModel.FindOne(l.ctx, uid)
 	encryptedPassword := EncryptPassword(req.NewPassword)
 	if encryptedPassword != user.Password {
 		// 可以更改密码
